Updated Schedule class
Firstly, it takes a Details instance instead of a value / id
Secondly, it's no longer generic. The Bakalari parser will now return this new type
And lastly, the merge function isn't static anymore. This is because of the previous change

Unrelated, I decided not to do a tuple for the period types because it doesn't always make sense. Now it's just an array